feat: Add MarkdownSearchDebugger for search debugging and logging
